feat: Added react router (no new functionality)
- Removed ConversationProvider

- Added react routing to navigate across
chats, now router maintains the state of 'current conversation'

- Fixed
annoying layout issues due to h-full and w-full on high level elements

 Little tricky to use the react router navigation after updating
state. After creating a new conversation for example, and I need to
fetch the new list of conversations, we get some bugs when navigating to
the conversation page (since that conversation still doesn't exist in
the list of conversations due to delayed state update before
navigating). Way to get around this: Try not to? Please? But if you
really have to: Added the conversations list along to the params as
dependencies in the useEffect of the Conversation component. This means
that we'll still get that 'buggy' navigation but it will self correct
shortly thereafter.